Move TimespanDatabaseRepresentation to its own file.
