+++ python/lsst/daf/butler/dimensions/_records.py
@@ -37,7 +37,7 @@ from lsst.daf.butler._compat import PYDANTIC_V2, _BaseModelCompat
 from lsst.utils.classes import immutable
 from pydantic import Field, StrictBool, StrictFloat, StrictInt, StrictStr, create_model
 
-from .._timespan import Timespan, TimespanDatabaseRepresentation
+from .._timespan import Timespan
 from ..json import from_json_pydantic, to_json_pydantic
 from ..persistence_context import PersistenceContextVars
 from ._elements import Dimension, DimensionElement
@@ -69,7 +69,7 @@ def _subclassDimensionRecord(definition: DimensionElement) -> type[DimensionReco
     if definition.spatial:
         slots.append("region")
     if definition.temporal:
-        slots.append(TimespanDatabaseRepresentation.NAME)
+        slots.append("timespan")
     d = {"definition": definition, "__slots__": tuple(slots), "fields": fields}
     return type(definition.name + ".RecordClass", (DimensionRecord,), d)
