+++ python/lsst/daf/butler/registry/dimensions/table.py
@@ -41,7 +41,6 @@ from lsst.daf.relation import Join, Relation, sql
 from ..._column_tags import DimensionKeyColumnTag
 from ..._column_type_info import LogicalColumn
 from ..._named import NamedKeyMapping
-from ..._timespan import TimespanDatabaseRepresentation
 from ...dimensions import (
     DatabaseDimensionElement,
     DataCoordinate,
@@ -51,6 +50,7 @@ from ...dimensions import (
     SkyPixDimension,
     addDimensionForeignKey,
 )
+from ...timespan_database_representation import TimespanDatabaseRepresentation
 from .. import queries
 from ..interfaces import (
     Database,
